UML Class diagrams
Created uml class diagrams. 
In the root i have "App" java class and four packages: "database", "models", "ui" and "utils". 
There are two classes "DatabaseConfigurator" and "RealmManager" in the "database" package. These classes designed for connection and work with Realm darabase.
In the "models" package I have classes that need to work with the information stored in the database. Each class presents to a separate table and fields of class presents to a separate columns in the tables.
In the "ui" package I have classes for activity and fragments, and package for adapters.
In the "utils" package I have two classes: DebugLogger for writing logs in console; RecyclerViewUtil for showing information using RecyclerView.